integrated AddDrop Modal
