+++ .gitignore
@@ -34,9 +34,7 @@ local.properties
 # node.js
 #
 node_modules/
-npm-debug.log
-yarn-error.log
-yarn.lock
+*.lock
 
 # BUCK
 buck-out/
